Replace cost amount int with float64 type
When the user has Euro currency set for their account.

* dnsdomains: Amount int to float64
* emaildomains: Amount int to float64
* ips: Amount int to float64
* loadbalancers: Amount int to float64

+++ v2/dnsdomains.go
@@ -30,9 +30,9 @@ type DNSDomain struct {
 
 // DNSDomainPrice represents the price for a single domain
 type DNSDomainPrice struct {
-	Amount   int    `json:"amount"`
-	Currency string `json:"currency"`
-	Years    int    `json:"years"`
+	Amount   float64 `json:"amount"`
+	Currency string  `json:"currency"`
+	Years    int     `json:"years"`
 }
 
 // AddDNSDomainParams - used for adding existing domains to GleSYS
